% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_properties.R
\name{add_properties}
\alias{add_bbox}
\alias{add_crs}
\alias{add_properties}
\title{Add properties, crs, and bounding box}
\usage{
add_properties(x, ...)

add_crs(x, crs)

add_bbox(x, bbox = NULL)
}
\arguments{
\item{x}{An object of class \code{geojson}}

\item{...}{Properties to be added, supports NSE as well as SE}

\item{crs}{(character) a CRS string. required.}

\item{bbox}{(numeric) a vector or list of length 4. If \code{NULL},
the bounding box is calculated for you}
}
\description{
Add properties, crs, and bounding box
}
\examples{
# add properties
x <- '{ "type": "LineString", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ] }'
(y <- linestring(x))
y \%>\% feature() \%>\% add_properties(population = 1000)

# add crs
crs <- '{"type": "name",
 "properties": {
     "name": "urn:ogc:def:crs:OGC:1.3:CRS84"
}}'
y \%>\% feature() \%>\% add_crs(crs)

# add bbox
x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0] ]
  ]
}'
(y <- polygon(x))
y \%>\% feature() \%>\% add_bbox()
y \%>\% feature() \%>\% add_bbox(c(-10.0, -10.0, 10.0, 10.0))
}

