% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.geojson.R, R/sp_to_geojson.R
\docType{methods}
\name{as.geojson}
\alias{as.geojson}
\alias{as.geojson,SpatialLines-method}
\alias{as.geojson,SpatialLinesDataFrame-method}
\alias{as.geojson,SpatialPoints-method}
\alias{as.geojson,SpatialPointsDataFrame-method}
\alias{as.geojson,SpatialPolygons-method}
\alias{as.geojson,SpatialPolygonsDataFrame-method}
\alias{as.geojson,character-method}
\alias{as.geojson,json-method}
\title{Geojson class}
\usage{
as.geojson(x)

\S4method{as.geojson}{json}(x)

\S4method{as.geojson}{character}(x)

\S4method{as.geojson}{SpatialPointsDataFrame}(x)

\S4method{as.geojson}{SpatialPoints}(x)

\S4method{as.geojson}{SpatialLinesDataFrame}(x)

\S4method{as.geojson}{SpatialLines}(x)

\S4method{as.geojson}{SpatialPolygonsDataFrame}(x)

\S4method{as.geojson}{SpatialPolygons}(x)
}
\arguments{
\item{x}{input}
}
\description{
Geojson class
}
\examples{
as.geojson(geojson_data$featurecollection_point)

x <- geojson_data$polygons_average
as.geojson(x)

x <- geojson_data$polygons_aggregate
as.geojson(x)

x <- geojson_data$points_count
as.geojson(x)
}

