% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\alias{bbox_add}
\alias{bbox_get}
\title{Add or get bounding box}
\usage{
bbox_add(x, bbox = NULL)

bbox_get(x)
}
\arguments{
\item{x}{An object of class \code{geojson}}

\item{bbox}{(numeric) a vector or list of length 4. If \code{NULL},
the bounding box is calculated for you}
}
\description{
Add or get bounding box
}
\details{
Note that \code{bbox_get} outputs the bbox if it exists, but
does not calculate it from the geojson. See \code{\link{geo_bbox}}
to calculate a bounding box
}
\examples{
x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0] ]
  ]
}'

# add bbox
(y <- polygon(x))
y \%>\% feature() \%>\% bbox_add()
y \%>\% feature() \%>\% bbox_add(c(-10.0, -10.0, 10.0, 10.0))

# get bounding box
z <- y \%>\% feature() \%>\% bbox_add()
bbox_get(z)

## returns NULL if no bounding box
bbox_get(x)
}
\references{
\url{http://geojson.org/geojson-spec.html#bounding-boxes}
}

