% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_box.R
\name{bound_box}
\alias{bound_box}
\title{Bounding box}
\usage{
bound_box(x)
}
\arguments{
\item{x}{input}
}
\description{
Bounding box
}
\examples{
library("lawn")
bound_box(as.geojson(lawn_data$polygons_count))
}

