% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature}
\alias{feature}
\title{feature class}
\usage{
feature(x)
}
\arguments{
\item{x}{input}
}
\description{
feature class
}
\examples{
x <- '{ "type": "Point", "coordinates": [100.0, 0.0] }'
(y <- feature(x))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)

# make multilinestring into a feature
x <- '{ "type": "MultiLineString", "coordinates": [ [ [100.0, 0.0], [101.0, 1.0] ], [ [102.0, 2.0], [103.0, 3.0] ] ] }'
(y <- multilinestring(x))
feature(y)
}

