% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurecollection.R
\name{featurecollection}
\alias{featurecollection}
\title{featurecollection class}
\usage{
featurecollection(x)
}
\arguments{
\item{x}{input}
}
\description{
featurecollection class
}
\examples{
file <- system.file("examples", 'featurecollection1.geojson', package = "geojson")
file <- system.file("examples", 'featurecollection2.geojson', package = "geojson")
str <- paste0(readLines(file), collapse = " ")
(y <- featurecollection(str))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)
}

