% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_pretty.R
\name{geo_pretty}
\alias{geo_pretty}
\title{Pretty print geojson}
\usage{
geo_pretty(x)
}
\arguments{
\item{x}{input, an object of class \code{geojson}}
}
\description{
Pretty print geojson
}
\details{
Wrapper around \code{\link[jsonlite]{prettify}}
}
\examples{
geo_pretty(point('{ "type": "Point", "coordinates": [100.0, 0.0] }'))

x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [100.0, 1.0], [101.0, 1.0], [101.0, 0.0], [100.0, 0.0] ]
  ]
}'
poly <- polygon(x)
geo_pretty(poly)
}
