% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson-methods.R
\name{geo_type}
\alias{geo_pretty}
\alias{geo_type}
\alias{geo_write}
\title{convenience methods to get various geojson attribute information}
\usage{
geo_type(x)

geo_pretty(x)

geo_write(x, file)
}
\arguments{
\item{x}{input}

\item{file}{a file name}
}
\description{
convenience methods to get various geojson attribute information
}

