% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_write.R
\name{geo_write}
\alias{geo_write}
\title{Write geojson to disk}
\usage{
geo_write(x, file)
}
\arguments{
\item{x}{input, an object of class \code{geojson}}

\item{file}{(character) a file path, or connection}
}
\description{
Write geojson to disk
}
\details{
Wrapper around \code{\link[jsonlite]{toJSON}} and
\code{\link{cat}}
}
\examples{
file <- tempfile(fileext = ".geojson")
geo_write(
  point('{ "type": "Point", "coordinates": [100.0, 0.0] }'),
  file
)
readLines(file)
unlink(file)
}
