% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring.R
\name{linestring}
\alias{linestring}
\title{linestring class}
\usage{
linestring(x)
}
\arguments{
\item{x}{input}
}
\description{
linestring class
}
\examples{
x <- '{ "type": "LineString", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ] }'
(y <- linestring(x))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)
}

