% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipoint.R
\name{multipoint}
\alias{multipoint}
\title{multipoint class}
\usage{
multipoint(x)
}
\arguments{
\item{x}{input, either JSON character string or a list}
}
\value{
an object of class geomultipoint and either geojson (string input),
or geo_list (list input)
}
\description{
multipoint class
}
\examples{
x <- '{"type": "MultiPoint", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ] }'
(y <- multipoint(x))
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))

# list
x <- list(type = "MultiPoint", coordinates =
  matrix(c(100.0, 101.0, 0.0, 1.0), ncol = 2))
(y <- multipoint(x))
y
}
