% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipolygon.R
\name{multipolygon}
\alias{multipolygon}
\title{multipolygon class}
\usage{
multipolygon(x)
}
\arguments{
\item{x}{input, either JSON character string or a list}
}
\value{
an object of class geomultipolygon and either geojson (string input),
or geo_list (list input)
}
\description{
multipolygon class
}
\examples{
x <- '{ "type": "MultiPolygon",
"coordinates": [
  [[[102.0, 2.0], [103.0, 2.0], [103.0, 3.0], [102.0, 3.0], [102.0, 2.0]]],
  [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]],
  [[100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2]]]
  ]
}'
(y <- multipolygon(x))
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))

# list
x <- list(type = "MultiPolygon")
mat1 <- list(
  matrix(c(100, 101, 0, 1), ncol = 2),
  matrix(c(102, 103, 1, 2), ncol = 2),
  matrix(c(104, 105, 2, 3), ncol = 2)
)
mat2 <- lapply(mat1, function(z) z + 5)
mat3 <- lapply(mat1, function(z) z - 7)
x$coordinates <- list(mat1, mat2, mat3)
(y <- multipolygon(x))
}
