% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Search geojson}
\usage{
overlay(points, polygons)
}
\description{
Search geojson
}
\examples{
x <- "{\\"type\\":\\"FeatureCollection\\",\\"features\\":[{\\"type\\":\\"Feature\\",\\"geometry\\":{\\"type\\":\\"Point\\",\\"coordinates\\":[-99.74,32.45]},\\"properties\\":{}}]}"
x <- as.geojson(x)
library("lawn")
lawn_data$points_within \%>\% view
overlay(lawn_data$points_within, lawn_data$polygons_within) \%>\% view
}

