% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\title{polygon class}
\usage{
polygon(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{(ignored in geojson use, extra arguments passed to \code{\link[graphics]{polygon}})}
}
\description{
polygon class
}
\section{Note}{

This function masks the non-generic graphics package function of the same
name, if the input is interpretable by \code{\link[grDevices]{xy.coords}}
then the default method will hand back to \code{\link[graphics]{polygon}}.
}
\examples{
x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [100.0, 1.0], [101.0, 1.0], [101.0, 0.0], [100.0, 0.0] ]
  ]
}'
(y <- polygon(x))
y[1]
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0] ],
  [ [100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2] ]
  ]
}'
(y <- polygon(x))
y[1]
geo_type(y)
geo_pretty(y)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))
}
