% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift_client.R
\name{sift_client}
\alias{sift_client}
\alias{sift_client.character}
\alias{sift_client.geojson}
\alias{sift_client.json}
\alias{sift_client.list}
\title{Sift geojson - low level client}
\usage{
sift_client(x, query)

\method{sift_client}{character}(x, query)

\method{sift_client}{json}(x, query)

\method{sift_client}{geojson}(x, query)

\method{sift_client}{list}(x, query)
}
\arguments{
\item{x}{input, one of character string, json, list, or ...}

\item{query}{query}
}
\value{
an object of class \code{character}
}
\description{
Sift geojson - low level client
}
\examples{
library("leaflet")

# get sample data
file <- system.file("examples", "zillow_or.geojson", package = "siftgeojson")

# plot as is
dat <- jsonlite::fromJSON(file, FALSE)
leaflet() \%>\% addTiles() \%>\% addGeoJSON(dat) \%>\% setView(-122.8, 44.8, zoom = 8)

# filter to features in Multnomah County only
json <- paste0(readLines(file), collapse = "")
json2 <- sift_client(json, '.features[] | select(.properties.COUNTY == "Multnomah")')
dat <- jsonlite::fromJSON(json2, FALSE)
leaflet() \%>\% addTiles() \%>\% addGeoJSON(dat) \%>\% setView(-122.6, 45.5, zoom = 10)

x <- "{\\"type\\":\\"FeatureCollection\\",\\"features\\":[{\\"type\\":\\"Feature\\",\\"geometry\\":{\\"type\\":\\"Point\\",\\"coordinates\\":[-99.74,32.45]},\\"properties\\":{}}]}"
x <- as.geojson(x)
sift_client(x, ".")
sift_client(x, "keys")

library("lawn")
x <- lawn_data$polygons_average
x <- as.geojson(x)
sift_client(x, ".")
sift_client(x, "keys")
sift_client(x, ".[]")
sift_client(x, ".features")
sift_client(x, ".features[]")
sift_client(x, ".features[].type")
sift_client(x, ".features[].properties")
sift_client(x, ".features[].geometry[]")
}

