% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sifter.R
\name{sifter}
\alias{sifter}
\alias{sifter_}
\title{Sift geojson - high level DSL}
\usage{
sifter(.data, ...)

sifter_(.data, ..., .dots)
}
\arguments{
\item{.data}{input, one of character string, json, list, or ...}

\item{...}{dots}

\item{.dots}{dots}
}
\value{
an object of class \code{character}
}
\description{
Sift geojson - high level DSL
}
\examples{
library("leaflet")

# get sample data
file <- system.file("examples", "zillow_or.geojson", package = "siftgeojson")

# plot as is
dat <- jsonlite::fromJSON(file, FALSE)
leaflet() \%>\% addTiles() \%>\% addGeoJSON(dat) \%>\% setView(-122.8, 44.8, zoom = 8)

# filter to features in Multnomah County only
json <- paste0(readLines(file), collapse = "")
#json2 <- sifter(json, '.features[] | select(.properties.COUNTY == "Multnomah")')
#dat <- jsonlite::fromJSON(json2, FALSE)
#leaflet() \%>\% addTiles() \%>\% addGeoJSON(dat) \%>\% setView(-122.6, 45.5, zoom = 10)

# sift to Multnomah County only
res <- sifter(json, COUNTY == Multnomah)
## check that only Multnomah County came back
res \%>\% jqr::index() \%>\% jqr::dotstr(properties.COUNTY)
}

