% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.geojson.R
\name{to_geojson}
\alias{to_geojson}
\title{Convert GeoJSON character string to approriate GeoJSON class}
\usage{
to_geojson(x)
}
\arguments{
\item{x}{GeoJSON character string}
}
\description{
Automatically detects and adds the class
}
\examples{
mp <- '{"type":"MultiPoint","coordinates":[[100,0],[101,1]]}'
to_geojson(mp)

ft <- '{"type":"Feature","properties":{"a":"b"},
"geometry":{"type": "MultiPoint","coordinates": [ [100.0, 0.0], [101.0, 1.0] ]}}'
to_geojson(mp)

fc <- '{"type":"FeatureCollection","features":[{"type":"Feature","properties":{"a":"b"},
"geometry":{"type": "MultiPoint","coordinates": [ [100.0, 0.0], [101.0, 1.0] ]}}]}'
to_geojson(fc)
}
