% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.geojson.R
\name{as.geojson}
\alias{as.geojson}
\title{Geojson class}
\usage{
as.geojson(x)
}
\arguments{
\item{x}{input}
}
\description{
Geojson class
}
\examples{
as.geojson(geojson_data$featurecollection_point)

x <- geojson_data$polygons_average
as.geojson(x)

x <- geojson_data$polygons_aggregate
as.geojson(x)

x <- geojson_data$points_count
as.geojson(x)
}

