% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.geojson.R
\name{as.geojson}
\alias{as.geojson}
\title{Geojson class}
\usage{
as.geojson(x)
}
\description{
Geojson class
}
\examples{
x <- "{\\"type\\":\\"FeatureCollection\\",\\"features\\":[{\\"type\\":\\"Feature\\",\\"geometry\\":{\\"type\\":\\"Point\\",\\"coordinates\\":[-99.74,32.45]},\\"properties\\":{}}]}"
as.geojson(x)
}

