% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurecollection.R
\name{featurecollection}
\alias{featurecollection}
\title{featurecollection class}
\usage{
featurecollection(x)
}
\arguments{
\item{x}{input}
}
\description{
featurecollection class
}
\examples{
file <- system.file("examples", 'featurecollection1.geojson', package = "geojson")
file <- system.file("examples", 'featurecollection2.geojson', package = "geojson")
str <- paste0(readLines(file), collapse = " ")
(y <- featurecollection(str))
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))
}

