% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_type.R
\name{geo_type}
\alias{geo_type}
\title{Get geometry type}
\usage{
geo_type(x)
}
\arguments{
\item{x}{input, an object of class \code{geojson}}
}
\description{
Get geometry type
}
\examples{
geo_type(point('{ "type": "Point", "coordinates": [100.0, 0.0] }'))

x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [100.0, 1.0], [101.0, 1.0], [101.0, 0.0], [100.0, 0.0] ]
  ]
}'
poly <- polygon(x)

geo_type(poly)
}
