% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson-package.R
\docType{package}
\name{geojson-package}
\alias{geojson}
\alias{geojson-package}
\title{geojson}
\description{
Classes for GeoJSON to make working with GeoJSON easier
}
\author{
Scott Chamberlain, Jeroen Ooms
}
\keyword{package}

