% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring.R
\name{linestring}
\alias{linestring}
\title{linestring class}
\usage{
linestring(x)
}
\arguments{
\item{x}{input, either JSON character string or a list}
}
\value{
an object of class geolinestring and either geojson (string input),
or geo_list (list input)
}
\description{
linestring class
}
\examples{
x <- '{ "type": "LineString", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ]}'
(y <- linestring(x))
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))

# list
x <- list(type = "LineString")
x$coordinates <- matrix(c(100.0, 101.0, 0.0, 1.0), ncol = 2)
(y <- linestring(x))

lngs <- c(-131.5, -123.4, -124.5, -118.8, -122, -114.6, -108.3, -105.8,
  -105.1, -97.7, -96.3, -92.1, -90.7, -85.8, -84, -78, -69.6, -83.7, -90.4,
  -93.9, -100.9, -97.7, -117.8, -124.8, -132.9)
lats <- c(60.9, 60.4, 56.8, 55.4, 50.1, 51.4, 57.1, 55.2, 51.8, 55.6, 51.6,
  54, 50.7, 51.2, 48.9, 49.2, 55.2, 53.5, 56.4, 59.4, 63.9, 66.8, 65.9,
  70.5, 69.3)
x$coordinates <- matrix(c(lngs, lats), ncol = 2, byrow = FALSE)
(y <- linestring(x))
y \%>\% feature()
y \%>\% feature() \%>\% featurecollection()
}
