% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilinestring.R
\name{multilinestring}
\alias{multilinestring}
\title{multilinestring class}
\usage{
multilinestring(x)
}
\arguments{
\item{x}{input}
}
\description{
multilinestring class
}
\examples{
x <- '{ "type": "MultiLineString", "coordinates": [ [ [100.0, 0.0], [101.0, 1.0] ], [ [102.0, 2.0], [103.0, 3.0] ] ] }'
(y <- multilinestring(x))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)
}

