% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipoint.R
\name{multipoint}
\alias{multipoint}
\title{multipoint class}
\usage{
multipoint(x)
}
\arguments{
\item{x}{input}
}
\description{
multipoint class
}
\examples{
x <- '{"type": "MultiPoint", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ] }'
(y <- multipoint(x))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)
}

