% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Overlay}
\usage{
overlay(points, polygons)
}
\arguments{
\item{points}{geojson points}

\item{polygons}{geojson polygons}
}
\description{
Overlay
}
\examples{
library("lawn")
lawn_data$points_within \%>\% view
overlay(lawn_data$points_within, lawn_data$polygons_within) \%>\% view
}

