% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{point class}
\usage{
point(x)
}
\arguments{
\item{x}{input}
}
\description{
point class
}
\examples{
x <- '{ "type": "Point", "coordinates": [100.0, 0.0] }'
(y <- point(x))
y$string
y$type()
y$pretty()
y$write(file = (f <- tempfile(fileext = ".geojson")))
jsonlite::fromJSON(f, FALSE)
unlink(f)
}

