% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\title{polygon class}
\usage{
polygon(x)
}
\arguments{
\item{x}{input, either JSON character string or a list}
}
\value{
an object of class geopolygon and either geojson (string input),
or geo_list (list input)
}
\description{
polygon class
}
\examples{
x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [100.0, 1.0], [101.0, 1.0], [101.0, 0.0], [100.0, 0.0] ]
  ]
}'
(y <- polygon(x))
y[1]
geo_type(y)
geo_pretty(y)
geo_write(y, f <- tempfile(fileext = ".geojson"))
jsonlite::fromJSON(f, FALSE)
unlink(f)

x <- '{ "type": "Polygon",
"coordinates": [
  [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0] ],
  [ [100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2] ]
  ]
}'
(y <- polygon(x))
y[1]
geo_type(y)
geo_pretty(y)

# add to a data.frame
library('tibble')
data_frame(a = 1:5, b = list(y))

# list
x <- list(type = "Polygon")
x$coordinates <- list(
  matrix(c(100, 101, 0, 1), ncol = 2),
  matrix(c(102, 103, 1, 2), ncol = 2),
  matrix(c(104, 105, 2, 3), ncol = 2)
)
(y <- polygon(x))
}
